wasm: gl: render: bug-fix: doc:
 - wasm improvement
 - several bugs fix
 - debugging improvement
 - bloom effect improvement
 - job manager now ignore dangling jobs
 - readme now has link to wasm demos
 - several other needed functiuonalities

+++ gearoenix/render/reflection/gx-rnd-rfl-runtime.cpp
@@ -127,7 +127,7 @@ void gearoenix::render::reflection::Runtime::set_runtime_reflection_self(
                     self->name + "-radiance-target" + name_ext + "-" + std::to_string(mip_index),
                     std::vector {
                         texture::Attachment {
-                            .mipmap_level = static_cast<unsigned int>(mip_index),
+                            .mipmap_level = static_cast<std::uint8_t>(mip_index),
                             .var = texture::AttachmentCube {
                                 .txt = self->radiance,
                                 .face = face.face,
